backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "thtn-dev", email = "thtn.1611.dev@gmail.com" }]
requires-python = ">=3.12,<4.0"
readme = "README.md"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "google-genai (>=1.11.0,<2.0.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "langchain-google-genai (>=2.1.5,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "wheel (>=0.45.1,<0.46.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "google-auth (>=2.39.0,<3.0.0)",
    "orjson (>=3.10.18,<4.0.0)",
    "langchain-openai (>=0.3.19,<0.4.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiohttp (>=3.12.7,<4.0.0)",
]

[project.scripts]
start = "uvicorn app.main:app --host 127.0.0.1 --port 8000 --reload"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pylint>=3.3.7,<4",
    "ruff>=0.11.12",
]

[tool.uv]
package = false

[tool.pylint.MASTER]
ignore-paths = '^migrations/.*$'
disable = [
    'line-too-long',
    'trailing-newlines'
]

[tool.ruff]
# Target Python version
target-version = "py38"

# Line length
line-length = 88

# Exclude files/folders
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "migrations/",
]

[tool.ruff.lint]
# Enable rules
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings  
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]

# Disable specific rules
ignore = [
    "E501",  # line too long
    "B008",  # function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["myproject"]

[tool.ruff.format]
# Use single quotes
quote-style = "single"
